requirements.txt
django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
drf-spectacular==0.27.2
PyMySQL==1.1.1
gunicorn==21.2.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
